requirements.txt
# ProcEx — Procedural Cinematic Explainer Pipeline
# Core dependencies

# LLM SDKs (fallback chain: Claude → Gemini → OpenAI)
anthropic>=0.40.0
google-genai>=1.0.0
openai>=1.50.0

# TTS
elevenlabs>=1.0.0

# PDF parsing
pymupdf>=1.24.0         # fitz

# Image processing
Pillow>=10.0.0

# Manim (rendering engine — install separately if needed)
# pip install manim
# On some systems: sudo apt-get install texlive-latex-recommended

# Animation & video
# ffmpeg must be installed system-wide: apt-get install ffmpeg

# YAML skill packs
PyYAML>=6.0

# Utilities
numpy>=1.24.0

# ── Optional but recommended ──────────────────────────────────────────────────
# For async TTS in batch:
# aiohttp>=3.9.0

# System requirements (install via apt/brew, not pip):
#   ffmpeg           - video assembly, Ken Burns, concat
#   manim            - scene rendering (also: pip install manim)
#   texlive-latex    - required by Manim for MathTex equations
